Split reward-workspace out of its router (1097 -> 61 lines)
reward-interact was 1097 lines -- 206 of template and 890 of script --
because an entire encounter engine lived in the component whose only job is
choosing between browsing and working: tone controls, prompt preview,
session chat tracking, narrative turns and choices, a hero-image carousel.
All of it moved to reward-workspace.vue unchanged. Nothing about how a
Reward encounter works changed; the frame is what is shared, not the
contents.

SESSION STATE MOVED TO THE STORE, which is what made the split possible.
Interact mode is not just "a Reward is selected" -- a conversation already
in progress keeps you there after the Reward is deselected -- and that was
answered by a private `sessionChatIds = ref<number[]>([])` in the same file
that had to move. So the router would have lost sight of its own routing
condition. Session membership is a fact about the chats, so chatStore now
owns it, keyed by scope so Bots and Rewards do not share one session.
bot-interact has the identical knot and this is the answer for it too.

The status banner moved WITH the workspace rather than staying in the
router: every setStatus() call is a workspace action (prompt copied, story
returned, story continued), so it reports on that surface.

THE KIT COUNTER, FIXED TWICE. Splitting scenario dropped verifyNarrativeKit
to 6/7; splitting reward dropped it again. Both times the kit had moved one
hop down and a hand-kept INTERMEDIARIES list could not see it. Replaced with
a real graph walk -- and then had to fix the walk twice more:

  1. KIT_USE included kr-art-plate, which nearly every card renders, so
     under a transitive walk the count could never fall. A metric that
     cannot go down reports success for work nobody did. Now counts the
     conversation pieces only; the art plate's invariants are asserted
     directly by file, where they belong.
  2. The predicate tested the whole file, so
     `import type { NarrativeTurn } from '.../kr-chat-window.vue'` counted
     as adoption. Now requires the tag to open an element in the template.

Both found by mutation, not by reading. Stripping every kit element out of
reward-workspace while leaving the type import now correctly drops it to
6/7; before, it stayed a cheerful 7/7.

Rule 5: reward-interact 1097 -> 61 (-1036).

vue-tsc clean, eslint clean; narrative-kit, narrative-roles, route-gallery,
reachability, layout and wonderlab --strict all green.
