+++ components/rewards/reward-workspace.vue
@@ -1,0 +1,1105 @@
+<!-- /components/rewards/reward-workspace.vue -->
+<!--
+  The Reward working surface: an encounter with a chosen Reward.
+
+  Split out of reward-interact, which had grown to 1097 lines -- 206 of
+  template and 890 of script -- because a whole encounter engine lived in the
+  component whose only job is choosing between browsing and working. Tone
+  controls, prompt preview, session chat tracking, narrative turns and choices,
+  and a hero-image carousel all sit here now. dream-interact does the same
+  routing job in 57 lines because dream-workspace exists; this is that.
+
+  Silas, 2026-08-06, on the interact tier: it "will be inevitably less
+  consistent across models, since that's where we are actually hitting what we
+  do with them uniquely." So nothing about HOW a Reward encounter works changed
+  in the move -- the frame is what is shared, not the contents.
+
+  The status banner came along rather than staying in the router: every
+  setStatus() call is a workspace action (prompt copied, story returned, story
+  continued), so it reports on this surface and belongs to it.
+-->
+<template>
+  <section class="flex h-full min-h-0 w-full flex-col gap-4">
+    <!--
+      Was `v-else` against the static page-title hero that used to sit above it.
+      That hero duplicated the title workspace-header already renders from
+      frontmatter, so it is gone and this states its own condition instead: the
+      selected-reward bar belongs to interact mode.
+    -->
+    <header
+      v-if="isInteractMode"
+      class="flex shrink-0 items-center gap-3 rounded-2xl border border-base-300 bg-base-100 p-3 shadow-md"
+    >
+      <button
+        class="btn btn-ghost btn-sm shrink-0 rounded-xl"
+        type="button"
+        :disabled="isStarting"
+        @click="backToGallery"
+      >
+        <Icon name="kind-icon:arrow-left" class="h-4 w-4" />
+        <span class="hidden sm:inline">Rewards</span>
+      </button>
+
+      <div class="min-w-0 flex-1">
+        <h1 class="truncate text-lg font-bold text-primary md:text-xl">
+          {{ selectedRewardName }}
+        </h1>
+
+        <p class="hidden truncate text-xs text-base-content/50 sm:block">
+          Set the scene, then let the narrative gremlin loose.
+        </p>
+      </div>
+
+      <span
+        v-if="rewardStore.selectedReward"
+        class="badge shrink-0"
+        :class="rarityBadgeClass"
+      >
+        {{ rewardStore.selectedReward.rarity }}
+      </span>
+    </header>
+
+    <div
+      v-if="statusMessage"
+      class="shrink-0 rounded-2xl border p-3 text-sm"
+      :class="
+        statusTone === 'error'
+          ? 'border-error/40 bg-error/10 text-error'
+          : 'border-success/40 bg-success/10 text-success'
+      "
+    >
+      {{ statusMessage }}
+    </div>
+
+    <section
+      class="grid min-h-0 flex-1 grid-cols-1 gap-4 overflow-y-auto xl:grid-cols-[minmax(0,1fr)_minmax(320px,440px)] xl:overflow-hidden"
+    >
+      <div
+        class="grid min-h-[70vh] grid-rows-[auto_minmax(0,1fr)_auto] overflow-hidden rounded-2xl border border-base-300 bg-base-100 xl:min-h-0"
+      >
+        <div class="shrink-0 border-b border-base-300 p-4">
+          <article v-if="rewardStore.selectedReward">
+            <div class="flex flex-col gap-4 sm:flex-row">
+              <div
+                class="relative h-40 w-full shrink-0 overflow-hidden rounded-2xl border border-base-300 bg-base-300 sm:h-36 sm:w-36 md:h-44 md:w-44"
+              >
+                <img
+                  v-if="rewardHeroSrc"
+                  :src="rewardHeroSrc"
+                  :alt="selectedRewardName"
+                  class="h-full w-full object-cover"
+                  @error="handleHeroImageError"
+                />
+
+                <div
+                  v-else
+                  class="flex h-full w-full items-center justify-center bg-linear-to-br from-primary/10 to-base-300"
+                >
+                  <Icon
+                    :name="rewardStore.selectedReward.icon || 'kind-icon:gift'"
+                    class="h-16 w-16 text-primary/70"
+                  />
+                </div>
+              </div>
+
+              <div class="min-w-0 flex-1">
+                <p class="text-xl font-black text-primary">
+                  {{ selectedRewardName }}
+                </p>
+
+                <p
+                  v-if="rewardStore.selectedReward.flavorText"
+                  class="mt-1 text-sm italic text-base-content/60"
+                >
+                  "{{ rewardStore.selectedReward.flavorText }}"
+                </p>
+
+                <p
+                  v-if="rewardStore.selectedReward.description"
+                  class="mt-2 text-sm text-base-content/75"
+                >
+                  {{ rewardStore.selectedReward.description }}
+                </p>
+
+                <p
+                  v-if="rewardStore.selectedReward.effect"
+                  class="mt-2 text-sm text-base-content/75"
+                >
+                  <span class="font-bold text-secondary">Effect:</span>
+                  {{ rewardStore.selectedReward.effect }}
+                </p>
+
+                <div class="mt-3 flex flex-wrap gap-2">
+                  <span class="badge badge-primary badge-sm">
+                    {{ rewardStore.selectedReward.rewardType }}
+                  </span>
+
+                  <span
+                    v-if="rewardStore.selectedReward.collection"
+                    class="badge badge-outline badge-sm"
+                  >
+                    {{ rewardStore.selectedReward.collection }}
+                  </span>
+
+                  <span class="badge badge-sm" :class="rarityBadgeClass">
+                    {{ rewardStore.selectedReward.rarity }}
+                  </span>
+                </div>
+              </div>
+            </div>
+          </article>
+
+          <div
+            v-else
+            class="rounded-2xl border border-base-300 bg-base-200 p-4 text-sm text-base-content/60"
+          >
+            No reward selected. Head back to the gallery and pick something.
+          </div>
+        </div>
+
+        <!-- The reward story log is kr-chat-window (interface-vision Phase 5,
+             the second of the two remaining surfaces from t-051 --
+             character-interact.vue's PR #1400 is the reference shape). This
+             also removes the bespoke .story-dot keyframes below in favor of
+             kr-chat-window's own built-in streaming placeholder. -->
+        <kr-chat-window
+          class="bg-base-200 p-4"
+          :turns="chatTurns"
+          :label="`${selectedRewardName} story`"
+          :is-streaming="isStarting"
+          :streaming-text="streamingBotText"
+          streaming-label="The narrative gremlin is writing..."
+          empty-label="Start the encounter, then unleash the narrative gremlin."
+          :prose="false"
+          :selected-key="lastChoiceKey"
+          @choose="handleChoiceSelected"
+        >
+          <template v-if="canShowCustomFollowup" #footer>
+            <div class="rounded-2xl border border-base-300 bg-base-100 p-3">
+              <label class="form-control">
+                <span class="label py-1">
+                  <span
+                    class="label-text text-xs font-bold uppercase tracking-wide text-base-content/50"
+                  >
+                    Customize the next move
+                  </span>
+                </span>
+
+                <textarea
+                  v-model="customFollowup"
+                  class="textarea textarea-bordered min-h-20 rounded-2xl bg-base-200"
+                  placeholder="Take one of the paths, remix it, or do something wildly inadvisable..."
+                  :disabled="isStarting"
+                  @keydown.enter.exact.prevent="continueWithCustomPath"
+                />
+              </label>
+
+              <button
+                class="btn btn-accent mt-3 w-full rounded-2xl"
+                type="button"
+                :disabled="!canContinueCustom"
+                @click="continueWithCustomPath"
+              >
+                <span
+                  v-if="isStarting"
+                  class="loading loading-spinner loading-sm"
+                />
+                <Icon v-else name="kind-icon:wand" class="h-5 w-5" />
+                Continue Custom Path
+              </button>
+            </div>
+          </template>
+        </kr-chat-window>
+
+        <div class="shrink-0 border-t border-base-300 bg-base-100 p-3">
+          <div
+            class="flex flex-wrap items-center gap-2"
+            :class="sessionChats.length === 0 ? 'mb-2' : ''"
+          >
+            <button
+              class="btn btn-xs btn-ghost rounded-xl"
+              type="button"
+              :disabled="isStarting"
+              @click="clearPromptOptions"
+            >
+              Reset Options
+            </button>
+
+            <button
+              class="btn btn-xs btn-ghost rounded-xl"
+              type="button"
+              :disabled="isStarting"
+              @click="newStory"
+            >
+              New Story
+            </button>
+
+            <button
+              class="btn btn-xs btn-ghost rounded-xl"
+              type="button"
+              :disabled="!rewardPromptPreview"
+              @click="copyPrompt"
+            >
+              Copy Prompt
+            </button>
+
+            <span class="ml-auto text-xs text-base-content/50">
+              {{ activeServerName }}
+            </span>
+          </div>
+
+          <button
+            v-if="sessionChats.length === 0"
+            class="btn btn-success min-h-14 w-full rounded-2xl"
+            type="button"
+            :disabled="!canStartStory"
+            @click="startRewardStory"
+          >
+            <span
+              v-if="isStarting"
+              class="loading loading-spinner loading-sm"
+            />
+            <Icon v-else name="kind-icon:play" class="h-5 w-5" />
+            Start Story
+          </button>
+        </div>
+      </div>
+
+      <aside class="flex flex-col gap-4 xl:min-h-0 xl:overflow-y-auto xl:pr-1">
+        <EntityArtManager
+          v-if="rewardStore.selectedReward"
+          entity-type="reward"
+          :entity="rewardStore.selectedReward"
+          :slots="[
+            {
+              field: 'imagePath',
+              label: 'Reward image',
+              aspect: '1 / 1',
+              width: 1024,
+              height: 1024,
+            },
+            {
+              field: 'iconPath',
+              label: 'Icon',
+              aspect: '1 / 1',
+              width: 256,
+              height: 256,
+            },
+            {
+              field: 'cardPath',
+              label: 'Card',
+              aspect: '2 / 3',
+              width: 512,
+              height: 768,
+            },
+            {
+              field: 'heroPath',
+              label: 'Hero',
+              aspect: '16 / 9',
+              width: 1280,
+              height: 720,
+            },
+          ]"
+        />
+        <section
+          class="shrink-0 rounded-2xl border border-base-300 bg-base-100 p-4 shadow-md"
+        >
+          <h2 class="mb-3 text-lg font-bold text-base-content">
+            The Encounter
+          </h2>
+
+          <div class="grid grid-cols-1 gap-4">
+            <label class="form-control">
+              <span class="label">
+                <span class="label-text font-bold">How do you find it?</span>
+              </span>
+
+              <select
+                v-model="encounterMode"
+                class="select select-bordered bg-base-200"
+                :disabled="isStarting"
+              >
+                <option value="discover">
+                  You stumble across it unexpectedly
+                </option>
+                <option value="use">You deliberately use the item</option>
+                <option value="temptation">
+                  The item calls to you, and you can't ignore it
+                </option>
+                <option value="consequence">
+                  You already used it. Now deal with what happened.
+                </option>
+                <option value="custom">Custom scenario</option>
+              </select>
+            </label>
+
+            <label class="form-control">
+              <span class="label">
+                <span class="label-text font-bold">Tone</span>
+              </span>
+
+              <select
+                v-model="tone"
+                class="select select-bordered bg-base-200"
+                :disabled="isStarting"
+              >
+                <option value="whimsical">Whimsical</option>
+                <option value="dramatic">Dramatic</option>
+                <option value="ominous">Ominous</option>
+                <option value="funny">Funny</option>
+                <option value="heroic">Heroic</option>
+                <option value="weird">Deeply weird</option>
+              </select>
+            </label>
+          </div>
+
+          <label class="form-control mt-4">
+            <span class="label">
+              <span class="label-text font-bold">What do you do?</span>
+              <span class="label-text-alt text-base-content/50">Optional</span>
+            </span>
+
+            <textarea
+              v-model="customDirection"
+              class="textarea textarea-bordered min-h-24 rounded-2xl bg-base-200"
+              placeholder="Describe your action, setting, or whatever context feels relevant..."
+              :disabled="isStarting"
+            />
+          </label>
+        </section>
+
+        <section
+          class="shrink-0 rounded-2xl border border-base-300 bg-base-100 p-4 shadow-md"
+        >
+          <h2 class="mb-1 text-lg font-bold text-base-content">About You</h2>
+
+          <p class="mb-3 text-sm text-base-content/60">
+            Give the story engine some context about who's holding this thing.
+            This stays local to the prompt.
+          </p>
+
+          <label class="form-control">
+            <span class="label">
+              <span class="label-text font-bold">Your background</span>
+              <span class="label-text-alt text-base-content/50">Optional</span>
+            </span>
+
+            <textarea
+              v-model="userBackground"
+              class="textarea textarea-bordered min-h-24 rounded-2xl bg-base-200"
+              placeholder="A wandering cartographer with a bad knee and a knack for finding trouble. Currently lost somewhere between here and regret."
+              :disabled="isStarting"
+            />
+          </label>
+        </section>
+
+        <section
+          class="shrink-0 rounded-2xl border border-base-300 bg-base-100 p-4 shadow-md"
+        >
+          <button
+            class="flex w-full items-center justify-between gap-3 text-left"
+            type="button"
+            @click="showCharacterPanel = !showCharacterPanel"
+          >
+            <div class="min-w-0">
+              <h2 class="text-lg font-bold text-base-content">
+                Add a Character
+                <span class="badge badge-ghost badge-sm ml-2">Optional</span>
+              </h2>
+
+              <p class="mt-0.5 truncate text-sm text-base-content/60">
+                {{ selectedCharacterSummary }}
+              </p>
+            </div>
+
+            <Icon
+              :name="
+                showCharacterPanel
+                  ? 'kind-icon:chevron-up'
+                  : 'kind-icon:chevron-down'
+              "
+              class="h-5 w-5 shrink-0 text-base-content/50"
+            />
+          </button>
+
+          <div v-if="showCharacterPanel" class="mt-4">
+            <character-gallery
+              variant="dropdown"
+              title="Character"
+              subtitle="Optionally anchor the reward story to a character."
+              :show-controls="false"
+              :show-images="true"
+              :show-card-actions="false"
+              :show-meta="false"
+              :compact="true"
+            />
+
+            <button
+              v-if="characterStore.selectedCharacter"
+              class="btn btn-ghost btn-xs mt-3 rounded-xl"
+              type="button"
+              :disabled="isStarting"
+              @click="characterStore.deselectCharacter?.()"
+            >
+              <Icon name="kind-icon:x" class="h-4 w-4" />
+              Remove character
+            </button>
+          </div>
+        </section>
+
+        <section
+          class="shrink-0 rounded-2xl border border-base-300 bg-base-100 p-4 shadow-md"
+        >
+          <div class="flex items-center justify-between gap-3">
+            <button
+              class="flex min-w-0 flex-1 items-center justify-between gap-3 text-left"
+              type="button"
+              @click="showPromptPreview = !showPromptPreview"
+            >
+              <h2 class="text-lg font-bold text-base-content">
+                Prompt Preview
+              </h2>
+
+              <Icon
+                :name="
+                  showPromptPreview
+                    ? 'kind-icon:chevron-up'
+                    : 'kind-icon:chevron-down'
+                "
+                class="h-5 w-5 shrink-0 text-base-content/50"
+              />
+            </button>
+
+            <button
+              class="btn btn-xs btn-ghost shrink-0 rounded-xl"
+              type="button"
+              :disabled="!rewardPromptPreview"
+              @click="copyPrompt"
+            >
+              Copy
+            </button>
+          </div>
+
+          <pre
+            v-if="showPromptPreview"
+            class="mt-3 max-h-72 overflow-auto whitespace-pre-wrap rounded-2xl bg-base-200 p-3 text-xs text-base-content/70"
+            >{{ rewardPromptPreview }}</pre>
+        </section>
+      </aside>
+    </section>
+  </section>
+</template>
+
+<script setup lang="ts">
+import { computed, onMounted, ref, watch } from 'vue'
+import { useCharacterStore } from '@/stores/characterStore'
+import { useChatStore } from '@/stores/chatStore'
+import { useRewardStore } from '@/stores/rewardStore'
+import { useServerStore } from '@/stores/serverStore'
+import { useUserStore } from '@/stores/userStore'
+import type { ArtImage } from '@/stores/artStore'
+import type { Chat, Reward } from '~/prisma/generated/prisma/client'
+import type { NarrativeTurn } from '@/components/narrative/kr-chat-window.vue'
+import type { NarrativeChoice } from '@/components/narrative/kr-choice-list.vue'
+
+type ChatRuntimeInput = Parameters<
+  ReturnType<typeof useChatStore>['addChat']
+>[0]
+
+type RewardStoryMessage = {
+  role: 'system' | 'user' | 'assistant'
+  content: string
+}
+
+type ParsedChoice = {
+  key: string
+  label: string
+  text: string
+}
+
+type RewardWithArt = Reward & {
+  ArtImage?: ArtImage | null
+}
+
+const rewardStore = useRewardStore()
+const characterStore = useCharacterStore()
+const chatStore = useChatStore()
+const serverStore = useServerStore()
+const userStore = useUserStore()
+
+const selectedCharacterSummary = computed(() => {
+  const character = characterStore.selectedCharacter
+
+  if (!character) {
+    return 'No character attached. The reward may menace a generic silhouette.'
+  }
+
+  const species = character.species || 'unknown species'
+  const characterClass = character.class || 'unknown class'
+
+  return `${selectedCharacterTitle.value} · ${species} / ${characterClass}`
+})
+
+const encounterMode = ref<
+  'discover' | 'use' | 'temptation' | 'consequence' | 'custom'
+>('discover')
+const tone = ref('whimsical')
+const customDirection = ref('')
+const customFollowup = ref('')
+const userBackground = ref(userStore.user?.bio ?? '')
+const showCharacterPanel = ref(false)
+const showPromptPreview = ref(false)
+const statusMessage = ref('')
+const statusTone = ref<'success' | 'error'>('success')
+const isStarting = ref(false)
+/*
+ * Session membership lives in chatStore, keyed by scope, because the ROUTER
+ * needs it too: reward-interact stays in interact mode while a conversation is
+ * in progress, and a private ref here would have hidden that from it.
+ */
+const SESSION_SCOPE = 'reward'
+const heroImageIndex = ref(0)
+
+const selectedCharacterTitle = computed(() => {
+  const character = characterStore.selectedCharacter
+
+  if (!character) return 'No character selected'
+
+  return character.name
+    ? `${character.name} the ${character.honorific || 'Unremarkable'}`.trim()
+    : 'Unnamed character'
+})
+
+const activeServerName = computed(() => {
+  return (
+    serverStore.activeTextServer?.label ||
+    serverStore.activeTextServer?.title ||
+    'No text server selected'
+  )
+})
+
+const selectedRewardName = computed(() => {
+  return rewardStore.selectedReward?.name || 'the reward'
+})
+
+const rarityBadgeClass = computed(() => {
+  switch (rewardStore.selectedReward?.rarity) {
+    case 'UNCOMMON':
+      return 'badge-success'
+    case 'RARE':
+      return 'badge-info'
+    case 'EPIC':
+      return 'badge-secondary'
+    case 'LEGENDARY':
+      return 'badge-warning'
+    case 'MYTHIC':
+      return 'badge-error'
+    default:
+      return 'badge-ghost'
+  }
+})
+
+// Ordered fallback chain for the hero image: ArtImage base64 -> thumbnail ->
+// imagePath -> slug-derived /images/rewards/{type}/{slug}.webp. A failed load
+// advances to the next candidate instead of giving up.
+const heroImageCandidates = computed<string[]>(() => {
+  const reward = rewardStore.selectedReward as RewardWithArt | null
+
+  if (!reward) return []
+
+  const candidates: string[] = []
+  const image = reward.ArtImage
+
+  // Path-first: stored paths before inline base64; the <img> onerror handler
+  // advances through this list, so base64 stays a graceful fallback.
+  const artPath =
+    image?.imagePath?.trim() ||
+    (image as { path?: string | null } | null | undefined)?.path?.trim() ||
+    ''
+
+  if (artPath) candidates.push(artPath)
+
+  const path = reward.imagePath?.trim()
+
+  if (path) {
+    candidates.push(path)
+
+    if (
+      !path.startsWith('/') &&
+      !path.startsWith('http') &&
+      !path.startsWith('data:')
+    ) {
+      candidates.push(`/${path}`)
+    }
+  }
+
+  const slug = reward.slug?.trim()
+  const rewardType = reward.rewardType?.toLowerCase()
+
+  if (slug && rewardType) {
+    candidates.push(`/images/rewards/${rewardType}/${slug}.webp`)
+  }
+
+  if (image?.imageData) {
+    candidates.push(
+      `data:${normalizeImageMime(image.fileType)};base64,${image.imageData}`,
+    )
+  }
+
+  if (image?.thumbnailData) {
+    candidates.push(
+      `data:${normalizeImageMime(image.fileType)};base64,${image.thumbnailData}`,
+    )
+  }
+
+  return Array.from(new Set(candidates))
+})
+
+const rewardHeroSrc = computed(() => {
+  return heroImageCandidates.value[heroImageIndex.value] ?? ''
+})
+
+function handleHeroImageError() {
+  heroImageIndex.value += 1
+}
+
+function normalizeImageMime(fileType?: string | null) {
+  const fallback = 'image/webp'
+  const cleaned = fileType?.trim().replace(/^\./, '')
+
+  if (!cleaned) return fallback
+  if (cleaned.startsWith('image/')) return cleaned
+
+  return `image/${cleaned}`
+}
+
+const sessionChats = computed<Chat[]>(() => {
+  return chatStore.chats.filter((chat) =>
+    chatStore.sessionChatIds(SESSION_SCOPE).includes(chat.id),
+  )
+})
+
+const isInteractMode = computed(() => {
+  return Boolean(rewardStore.selectedReward) || sessionChats.value.length > 0
+})
+
+const latestSessionChat = computed(() => {
+  return sessionChats.value.at(-1) ?? null
+})
+
+const isResponding = computed(() => {
+  return sessionChats.value.some((chat) => !chat.botResponse)
+})
+
+const canStartStory = computed(() => {
+  return Boolean(
+    rewardStore.selectedReward && !isStarting.value && !isResponding.value,
+  )
+})
+
+const canContinueCustom = computed(() => {
+  return Boolean(
+    latestSessionChat.value?.botResponse &&
+    customFollowup.value.trim() &&
+    !isStarting.value &&
+    !isResponding.value,
+  )
+})
+
+const rewardPromptPreview = computed(() => buildRewardPrompt())
+
+const systemPrompt = computed(() => {
+  return [
+    'You are a narrative game master for Kind Robots.',
+    'Write vivid, interactive story scenes centered on strange rewards, artifacts, boons, curses, or plot devices.',
+    'Keep the scene immersive, concrete, and readable.',
+    'End every response with exactly 3 numbered follow-up choices.',
+    'Choice 1 must be cautious.',
+    'Choice 2 must be bold.',
+    'Choice 3 must be weird.',
+    'Each choice must be on its own line and start with 1., 2., or 3.',
+    'Do not explain the prompt. Write the scene directly.',
+  ].join('\n')
+})
+
+const fullSessionMessages = computed<RewardStoryMessage[]>(() => {
+  return sessionChats.value.flatMap((chat) => {
+    const messages: RewardStoryMessage[] = [
+      {
+        role: 'user',
+        content: chat.content,
+      },
+    ]
+
+    if (chat.botResponse) {
+      messages.push({
+        role: 'assistant',
+        content: chat.botResponse,
+      })
+    }
+
+    return messages
+  })
+})
+
+function setStatus(messageText: string, tone: 'success' | 'error' = 'success') {
+  statusMessage.value = messageText
+  statusTone.value = tone
+}
+
+function isLatestRespondedChat(chatId: number) {
+  return Boolean(
+    latestSessionChat.value?.id === chatId &&
+    latestSessionChat.value.botResponse,
+  )
+}
+
+function getStoryChoices(text: string | null | undefined): ParsedChoice[] {
+  if (!text) return []
+
+  const lines = text.split(/\r?\n/)
+  const choices: ParsedChoice[] = []
+
+  for (const line of lines) {
+    const match = line.match(/^\s*(1|2|3)[.)]\s+(.+?)\s*$/)
+
+    if (!match) continue
+
+    const choiceNumber = match[1]
+    const choiceText = match[2]?.trim()
+
+    if (choiceNumber !== '1' && choiceNumber !== '2' && choiceNumber !== '3') {
+      continue
+    }
+
+    if (!choiceText) continue
+
+    const label =
+      choiceNumber === '1'
+        ? 'Cautious'
+        : choiceNumber === '2'
+          ? 'Bold'
+          : 'Weird'
+
+    choices.push({
+      key: `${choiceNumber}-${choiceText}`,
+      label,
+      text: choiceText,
+    })
+  }
+
+  return choices.slice(-3)
+}
+
+function getStoryScene(text: string | null | undefined) {
+  if (!text) return ''
+
+  const lines = text.split(/\r?\n/)
+  const sceneLines = lines.filter((line) => {
+    return !/^\s*(1|2|3)[.)]\s+/.test(line)
+  })
+
+  return sceneLines.join('\n').trim()
+}
+
+function getPlayerDisplayText(chat: Chat, index: number) {
+  if (index === 0 && chat.content.includes('Reward:')) {
+    return `Begin the story with ${selectedRewardName.value}.`
+  }
+
+  if (chat.content.startsWith('Selected path:')) {
+    return chat.content.replace(/^Selected path:\s*/i, '').trim()
+  }
+
+  return chat.content
+}
+
+/* Each session Chat becomes a user turn plus (once it has a response) a
+   narrator turn. The in-flight chat is deliberately withheld from `turns`
+   while it is still streaming -- its growing text surfaces through
+   kr-chat-window's own `streamingText` placeholder instead, matching the
+   built-in streaming contract rather than mutating a turn's text in place. */
+const chatTurns = computed<NarrativeTurn[]>(() => {
+  const chats = sessionChats.value
+  const turns: NarrativeTurn[] = []
+
+  chats.forEach((chat, index) => {
+    turns.push({
+      id: `user-${chat.id}`,
+      text: getPlayerDisplayText(chat, index),
+      from: 'user',
+    })
+
+    const inFlight = isStarting.value && latestSessionChat.value?.id === chat.id
+    if (!chat.botResponse || inFlight) return
+
+    const choices: NarrativeChoice[] = getStoryChoices(chat.botResponse).map(
+      (choice) => ({
+        key: choice.key,
+        label: choice.label,
+        hint: choice.text,
+        disabled: !isLatestRespondedChat(chat.id),
+      }),
+    )
+
+    turns.push({
+      id: `narrator-${chat.id}`,
+      text: getStoryScene(chat.botResponse),
+      from: 'narrator',
+      speaker: selectedRewardName.value,
+      choices: choices.length ? choices : undefined,
+    })
+  })
+
+  return turns
+})
+
+const streamingBotText = computed(() => {
+  if (!isStarting.value) return ''
+  const chat = latestSessionChat.value
+  return chat?.botResponse ? getStoryScene(chat.botResponse) : ''
+})
+
+// Same visibility rule the old inline textarea used: once the latest chat has
+// any response text, whether or not it has finished streaming.
+const canShowCustomFollowup = computed(() =>
+  Boolean(latestSessionChat.value?.botResponse),
+)
+
+/*
+ * The choice most recently picked, so kr-chat-window can ring it (t-090).
+ * This surface had that highlight before its migration onto the shared kit and
+ * lost it, because kr-chat-window had no way to forward a selectedKey to the
+ * choice lists it embeds. The key is the display key kr-choice-list emits, not
+ * the recovered path text, since that is what it compares against.
+ */
+const lastChoiceKey = ref<string | null>(null)
+
+function handleChoiceSelected(choice: NarrativeChoice) {
+  lastChoiceKey.value = choice.key
+  // getStoryChoices keys are `${1|2|3}-${text}`; recover the raw path text
+  // rather than repurposing the display label/hint, which are free to change.
+  continueWithPath(choice.key.replace(/^[1-3]-/, ''))
+}
+
+function buildMessagesForRewardResponse(): RewardStoryMessage[] {
+  return [
+    {
+      role: 'system',
+      content: systemPrompt.value,
+    },
+    ...fullSessionMessages.value,
+  ]
+}
+
+function buildRewardPrompt() {
+  const reward = rewardStore.selectedReward
+
+  if (!reward) return ''
+
+  const character = characterStore.selectedCharacter
+  const direction = customDirection.value.trim()
+  const background = userBackground.value.trim()
+
+  const characterLine = character
+    ? `Character: ${character.name || 'Unnamed'} the ${character.honorific || 'Unremarkable'}. Species: ${character.species || 'Unknown'}, Class: ${character.class || 'Unknown'}, Personality: ${character.personality || 'unknown'}.`
+    : null
+
+  const lines = [
+    `Reward: ${reward.name}`,
+    reward.description ? `Description: ${reward.description}` : null,
+    reward.flavorText ? `Flavor text: ${reward.flavorText}` : null,
+    `Effect: ${reward.effect || 'No effect described.'}`,
+    `Type: ${reward.rewardType} · Collection: ${reward.collection || 'general'} · Rarity: ${reward.rarity}`,
+    background ? `Finder background: ${background}` : null,
+    characterLine,
+    `Encounter mode: ${encounterMode.value}`,
+    `Tone: ${tone.value}`,
+    direction ? `Player direction: ${direction}` : null,
+    '',
+    'Write a scene where this reward is discovered, earned, bonded with, invoked, or used.',
+    'Focus on what it feels like to encounter it, what it changes, and the immediate consequences.',
+    'End with exactly 3 numbered choices.',
+    '1. A cautious path.',
+    '2. A bold path.',
+    '3. A weird path.',
+  ]
+
+  return lines.filter((line): line is string => Boolean(line)).join('\n')
+}
+
+function buildFollowupPrompt(path: string) {
+  return [
+    `Selected path: ${path}`,
+    '',
+    'Continue the story from the selected path.',
+    'Respect the previous scene and consequences.',
+    'Advance the narrative with fresh action, sensory detail, and a meaningful cost or discovery.',
+    'End with exactly 3 numbered choices.',
+    '1. A cautious path.',
+    '2. A bold path.',
+    '3. A weird path.',
+  ].join('\n')
+}
+
+async function copyPrompt() {
+  const prompt = rewardPromptPreview.value
+
+  if (!prompt) return
+
+  await navigator.clipboard.writeText(prompt)
+  setStatus('Prompt copied.')
+}
+
+async function createRewardChat(content: string) {
+  const reward = rewardStore.selectedReward
+
+  if (!reward) {
+    throw new Error('Pick a reward before starting the story.')
+  }
+
+  const character = characterStore.selectedCharacter
+  const senderName = character?.name ?? userStore.user?.username ?? 'Anonymous'
+
+  const payload: ChatRuntimeInput = {
+    content,
+    type: 'Reward',
+    sender: senderName,
+    userId:
+      userStore.userId ??
+      userStore.user?.id ??
+      character?.userId ??
+      reward.userId ??
+      10,
+    characterId: character?.id ?? null,
+    recipientId: null,
+    serverId: serverStore.activeTextServer?.id ?? null,
+    isPublic: false,
+  }
+
+  const newChat = await chatStore.addChat(payload)
+
+  if (!newChat?.id) {
+    throw new Error('Failed to create reward story chat.')
+  }
+
+  chatStore.addSessionChat(SESSION_SCOPE, newChat.id)
+  chatStore.selectedChat = newChat
+
+  return newChat
+}
+
+async function streamRewardChat(chatId: number) {
+  if (typeof chatStore.streamResponse !== 'function') {
+    throw new Error('chatStore.streamResponse is not available.')
+  }
+
+  await chatStore.streamResponse(chatId, {
+    model: serverStore.activeTextServer?.model || 'gpt-4o-mini',
+    temperature: 0.85,
+    maxTokens: 2048,
+    serverId: serverStore.activeTextServer?.id ?? null,
+    stream: true,
+    messages: buildMessagesForRewardResponse(),
+  })
+}
+
+async function startRewardStory() {
+  const content = buildRewardPrompt()
+
+  if (!content.trim()) {
+    setStatus('Could not build a reward prompt.', 'error')
+    return
+  }
+
+  isStarting.value = true
+  statusMessage.value = ''
+  customFollowup.value = ''
+
+  try {
+    const newChat = await createRewardChat(content)
+    await streamRewardChat(newChat.id)
+    setStatus('Reward story returned.')
+  } catch (error) {
+    console.error('Error starting reward story:', error)
+    setStatus(
+      error instanceof Error ? error.message : 'Error starting reward story.',
+      'error',
+    )
+  } finally {
+    isStarting.value = false
+  }
+}
+
+async function continueWithPath(path: string) {
+  if (!path.trim() || isStarting.value) return
+
+  customFollowup.value = path
+
+  await continueRewardStory(path)
+}
+
+async function continueWithCustomPath() {
+  const path = customFollowup.value.trim()
+
+  if (!path || isStarting.value) return
+
+  await continueRewardStory(path)
+}
+
+async function continueRewardStory(path: string) {
+  isStarting.value = true
+  statusMessage.value = ''
+
+  try {
+    const newChat = await createRewardChat(buildFollowupPrompt(path))
+    customFollowup.value = ''
+    await streamRewardChat(newChat.id)
+    setStatus('Story continued.')
+  } catch (error) {
+    console.error('Error continuing reward story:', error)
+    setStatus(
+      error instanceof Error ? error.message : 'Error continuing reward story.',
+      'error',
+    )
+  } finally {
+    isStarting.value = false
+  }
+}
+
+function newStory() {
+  chatStore.clearSessionChats(SESSION_SCOPE)
+  customFollowup.value = ''
+  statusMessage.value = ''
+}
+
+function backToGallery() {
+  newStory()
+  rewardStore.deselectReward()
+}
+
+function clearPromptOptions() {
+  encounterMode.value = 'discover'
+  tone.value = 'whimsical'
+  customDirection.value = ''
+  customFollowup.value = ''
+  userBackground.value = userStore.user?.bio ?? ''
+  statusMessage.value = ''
+}
+
+onMounted(async () => {
+  await Promise.all([
+    chatStore.initialize(),
+    ...(serverStore.hasLoaded
+      ? []
+      : [serverStore.initialize({ fetchRemote: true })]),
+  ])
+})
+
+watch(
+  () => rewardStore.selectedReward?.id,
+  () => {
+    heroImageIndex.value = 0
+  },
+)
+</script>
