+++ utils/scripts/verifyNarrativeKit.ts
@@ -165,8 +165,62 @@ const existing = SURFACES.filter((path) => {
     return false
   }
 })
-const KIT_USE =
-  /kr-chat-window|KrChatWindow|kr-choice-list|KrChoiceList|kr-art-plate|KrArtPlate/
+/*
+ * THE CONVERSATION PIECES, and deliberately not kr-art-plate.
+ *
+ * This used to include the art plate, which was harmless while adoption was
+ * checked one hop deep off a hand-kept list. Under a transitive walk it makes
+ * the number a lie: kr-art-plate is rendered by nearly every card in the app,
+ * so ANY surface that shows a card reaches it, and the count can never fall.
+ * A metric that cannot go down is worse than no metric -- it reports success
+ * for work nobody did.
+ *
+ * Caught by mutation: stripping kr-chat-window and kr-choice-list out of
+ * reward-workspace entirely left the count at a cheerful 7/7.
+ *
+ * The heading says "conversation kit", so this counts the conversation: the
+ * transcript, the choices, and the composer that takes a custom reply. The art
+ * plate's own invariants are asserted directly above by file, which is where
+ * they belong.
+ */
+const KIT_PIECES = [
+  'kr-chat-window',
+  'kr-choice-list',
+  'narrative-response-composer',
+]
+
+/**
+ * Does this source MOUNT a conversation piece?
+ *
+ * ELEMENT-LEVEL, IN THE TEMPLATE. This was a substring test over the whole
+ * file, which counted
+ *
+ *   import type { NarrativeTurn } from '@/components/narrative/kr-chat-window.vue'
+ *
+ * as adoption -- so a surface that imported a TYPE and hand-rolled everything
+ * else scored a tick. Caught by mutation: stripping every kit element out of
+ * reward-workspace left the count at 7/7 because the type import remained.
+ *
+ * That is the fourth time in one session that asserting on a mention rather
+ * than a use produced a false pass in this repo's verifiers. It is the house
+ * failure mode, and the fix is always the same: require the tag to open an
+ * element, in the template, after comments are stripped.
+ */
+function mountsKitPiece(source: string): boolean {
+  const template = source
+    .replace(/<script[\s\S]*?<\/script>/g, '')
+    .replace(/<style[\s\S]*?<\/style>/g, '')
+    .replace(/<!--[\s\S]*?-->/g, '')
+  return KIT_PIECES.some((piece) => {
+    const pascal = piece
+      .split('-')
+      .map((part) => part.charAt(0).toUpperCase() + part.slice(1))
+      .join('')
+    return new RegExp(
+      `<\\s*(?:Lazy|lazy-)?(?:${piece}|${pascal})(?=[\\s/>])`,
+    ).test(template)
+  })
+}
 
 /*
  * Count TRANSITIVE adoption, one level deep.
@@ -176,50 +230,122 @@ const KIT_USE =
  * the components were demonstrably live in both render paths -- understating
  * progress and, worse, hiding a regression if the stage stopped using them.
  */
-const INTERMEDIARIES = [
-  'components/narrative/kr-narrator-stage.vue',
-  // Both of these are themselves mounted by the product pages, and both now
-  // render kit pieces internally (the composer's suggested-response row is a
-  // kr-choice-list; the art status' rendered illustration is a kr-art-plate).
-  // A surface that mounts either is on the kit whether it says so or not.
-  'components/narrative/narrative-response-composer.vue',
-  'components/narrative/narrative-art-status.vue',
-  // Dreams reaches the kit through its slide-out dock rather than mounting the
-  // pieces itself. The dock's own chrome (card flip, musings, emoji bursts)
-  // stays local on purpose — it is genuinely Dreams', and folding it into the
-  // shared stage would bloat the piece every other surface imports.
-  'components/navigation/workspace-narrator.vue',
-  // The workspace half of the interact frame. Scenario's configure/play phases
-  // moved out of scenario-interact into scenario-workspace, taking the chat
-  // window and choice list with them -- so the surface still renders the kit,
-  // one hop further down. Without this the counter reports scenario as a
-  // holdout the moment its router gets thinner, which is backwards: the split
-  // is the improvement.
-  //
-  // Add each `<model>-workspace.vue` here as the remaining interacts are
-  // split, or better, replace this hand-kept list with a real graph walk --
-  // the same lesson verifyRouteGalleryContract.ts had to learn about
-  // filename-shaped assumptions.
-  'components/scenarios/scenario-workspace.vue',
-]
-const liveIntermediaries = INTERMEDIARIES.filter((path) =>
-  KIT_USE.test(read(path)),
-)
+/*
+ * ADOPTION IS FOUND BY WALKING, NOT BY A LIST.
+ *
+ * This used to be a hand-kept INTERMEDIARIES array -- kr-narrator-stage, the
+ * response composer, the art status, the Dreams dock -- matched one level deep.
+ * It broke twice in one afternoon: splitting scenario-workspace out of
+ * scenario-interact dropped the count to 6/7, and splitting reward-workspace
+ * out of reward-interact dropped it again. Both times the kit had not been
+ * abandoned; it had moved one hop further down, and a list of names could not
+ * see that.
+ *
+ * Reporting a refactor as a regression is the same filename-shaped blindness
+ * verifyRouteGalleryContract.ts had to learn out of. So this follows the actual
+ * component graph from each surface: a surface adopts the kit if anything it
+ * transitively renders does.
+ *
+ * ONE EXCLUSION. wonderlab-preview-host.vue mounts every component in the repo
+ * through import.meta.glob, so traversing into it would make every surface look
+ * adopted. Being exhibited in the museum is not being used.
+ */
+const MUSEUM_MOUNTS_EVERYTHING = 'wonderlab-preview-host'
 
+const SKIP_DIRS = new Set([
+  'node_modules',
+  '.nuxt',
+  '.git',
+  'dist',
+  '.output',
+  'abandonware',
+  'archives',
+  'cypress',
+  'sample',
+])
+
+function walkVue(directory: string, out: string[] = []): string[] {
+  let entries: string[]
+  try {
+    entries = readdirSync(directory, { withFileTypes: true }).map((e) => e.name)
+  } catch {
+    return out
+  }
+  for (const entry of entries) {
+    if (SKIP_DIRS.has(entry)) continue
+    const full = resolve(directory, entry)
+    if (entry.endsWith('.vue')) out.push(full)
+    else if (!entry.includes('.')) walkVue(full, out)
+  }
+  return out
+}
+
+const componentsByName = new Map<string, string>()
+for (const file of walkVue(resolve(root, 'components'))) {
+  const name = file.split('/').pop()?.replace('.vue', '') ?? ''
+  if (name && !componentsByName.has(name)) componentsByName.set(name, file)
+}
+
+const templateOf = (source: string): string =>
+  source
+    .replace(/<script[\s\S]*?<\/script>/g, '')
+    .replace(/<style[\s\S]*?<\/style>/g, '')
+    .replace(/<!--[\s\S]*?-->/g, '')
+
+const kebab = (tag: string): string =>
+  (/^[a-z]/.test(tag)
+    ? tag
+    : tag.replace(/(?<!^)(?=[A-Z])/g, '-').toLowerCase()
+  ).replace(/^lazy-/, '')
+
+const childCache = new Map<string, string[]>()
+function childComponents(name: string): string[] {
+  const cached = childCache.get(name)
+  if (cached) return cached
+  const file = componentsByName.get(name)
+  if (!file) {
+    childCache.set(name, [])
+    return []
+  }
+  const template = templateOf(readFileSync(file, 'utf8'))
+  const found = new Set<string>()
+  for (const [, tag] of template.matchAll(/<([A-Za-z][\w-]*)/g)) {
+    const key = kebab(tag ?? '')
+    if (
+      key !== name &&
+      key !== MUSEUM_MOUNTS_EVERYTHING &&
+      componentsByName.has(key)
+    ) {
+      found.add(key)
+    }
+  }
+  const result = [...found].sort()
+  childCache.set(name, result)
+  return result
+}
+
+/** Does this surface, or anything it renders, mount a kit piece? */
 function adopts(path: string): boolean {
-  const source = read(path)
-  if (KIT_USE.test(source)) return true
-  return liveIntermediaries.some((intermediary) => {
-    const name = intermediary.split('/').pop()?.replace('.vue', '') ?? ''
-    const pascal = name.replace(/(^|-)([a-z])/g, (_, __, c) => c.toUpperCase())
-    return new RegExp(`${name}|${pascal}|Lazy${pascal}`).test(source)
-  })
+  const seen = new Set<string>()
+  const stack = [path.split('/').pop()?.replace('.vue', '') ?? '']
+  while (stack.length) {
+    const current = stack.pop()
+    if (!current || seen.has(current)) continue
+    seen.add(current)
+    const file = componentsByName.get(current)
+    if (!file) continue
+    if (mountsKitPiece(readFileSync(file, 'utf8'))) return true
+    for (const child of childComponents(current)) {
+      if (!seen.has(child)) stack.push(child)
+    }
+  }
+  return false
 }
 
 const adopters = existing.filter(adopts)
 console.log(
   `\ninfo - conversation kit adopted by ${adopters.length}/${existing.length} surface(s) ` +
-    `(directly or via ${liveIntermediaries.length} shared stage)`,
+    '(directly, or through anything they render)',
 )
 for (const path of existing) {
   console.log(
